Revert "Optimisation et Nettoyage du code (#553)" (#683)
This reverts commit c308bdb1166f3f82594e8a8e45c3752f5867ca0d.

+++ src/main/java/fr/openmc/core/features/city/actions/CityChestAction.java
@@ -3,11 +3,14 @@ package fr.openmc.core.features.city.actions;
 import fr.openmc.core.features.city.City;
 import fr.openmc.core.features.city.conditions.CityChestConditions;
 import fr.openmc.core.utils.ItemUtils;
+import fr.openmc.core.items.CustomItemRegistry;
 import fr.openmc.core.utils.messages.MessageType;
 import fr.openmc.core.utils.messages.MessagesManager;
 import fr.openmc.core.utils.messages.Prefix;
 import net.kyori.adventure.text.Component;
 import org.bukkit.entity.Player;
+
+import java.util.Objects;
 
 import static fr.openmc.core.features.city.conditions.CityChestConditions.UPGRADE_PER_AYWENITE;
 import static fr.openmc.core.features.city.conditions.CityChestConditions.UPGRADE_PER_MONEY;
@@ -20,7 +23,7 @@ public class CityChestAction {
         int price = city.getChestPages() * UPGRADE_PER_MONEY;
         int aywenite = city.getChestPages() * UPGRADE_PER_AYWENITE;
 
-        city.updateBalance(-price);
+        city.updateBalance((double) -price);
 
         if (ItemUtils.takeAywenite(player, aywenite)) {
             city.saveChestContent(city.getChestPages() + 1, null);
